Lower pt threhold of jets.

+++ python/BristolNTuple_GenJets_cfi.py
@@ -4,7 +4,7 @@ nTupleGenJets = cms.EDProducer("BristolNTuple_GenJets",
     InputTag = cms.InputTag('slimmedGenJets'),
     Prefix = cms.string('GenJet.'),
     Suffix = cms.string(''),
-    minPt = cms.double(20),
+    minPt = cms.double(15),
     maxAbsoluteEta = cms.double(999),
     MaxSize = cms.uint32(99)
 )
